move: 댓글, 댓글 좋아요 코드 이동 (CommentController, CommentService 신규 생성)
