+++ activity-service/src/main/java/com/shop/activityservice/service/PostService.java
@@ -30,18 +30,6 @@ public class PostService {
         return postRepository.save(postWriteRequest.toEntity(writer));
     }
 
-    // 댓글 작성
-    @Transactional
-    public Comment writeComment(CommentWriteRequest commentWriteRequest, Long postId, String email) {
-        Post post = postRepository.findById(postId)
-                .orElseThrow(() -> new BaseException(ErrorCode.POST_NOT_FOUND));
-
-        User writer = userRepository.findByEmail(email)
-                .orElseThrow(() -> new BaseException(ErrorCode.USER_NOT_FOUND));
-
-        return commentRepository.save(commentWriteRequest.toEntity(post, writer));
-    }
-
     @Transactional
     public PostLike addPostLike(Long postId, String email) {
         Post post = postRepository.findById(postId)
@@ -63,24 +51,4 @@ public class PostService {
         return postLikeRepository.save(postLike);
     }
 
-    @Transactional
-    public CommentLike addCommentLike(Long commentId, String email) {
-        Comment comment = commentRepository.findById(commentId)
-                .orElseThrow(() -> new BaseException(ErrorCode.COMMENT_NOT_FOUND));
-
-        User writer = userRepository.findByEmail(email)
-                .orElseThrow(() -> new BaseException(ErrorCode.USER_NOT_FOUND));
-
-        boolean isCommentLike = commentLikeRepository.existsByCommentAndUser(comment, writer);
-        if (isCommentLike) {
-            throw new BaseException(ErrorCode.DUPLICATED_COMMENT_LIKE);
-        }
-
-        CommentLike commentLike = CommentLike.builder()
-                .comment(comment)
-                .user(writer)
-                .build();
-
-        return commentLikeRepository.save(commentLike);
-    }
 }
